optimized the way of adding or editing sections so that they only have to be changed in one place
