+++ src/components/resume-templates/basic/BasicDownload.tsx
@@ -14,7 +14,6 @@ type DocumentProps = {
 };
 
 const BasicDownload = ({ document }: DocumentProps) => {
-    console.log(document);
     const hyphenationCallback = (word: string) => {
         // Return word parts in an array
         return [word];
