+++ src/components/resume-templates/section-components/section-container-editor/SectionContainerEditor.module.css
@@ -2,26 +2,3 @@
     width: 100%;
     position: relative;
 }
-
-.addSectionButton {
-    position: absolute;
-    bottom: 0;
-    left: 50%;
-    transform: translate(-50%, 75%);
-    width: 50%;
-    height: 0.5em;
-    border-radius: 5px;
-    font-size: var(--font-size-medium);
-    background-color: #7467ff81;
-    color: white;
-    border: none;
-    cursor: pointer;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1;
-}
-
-.addSectionButton:hover {
-    background-color: #5f52e6;
-}
